@@ -1,14 +1,13 @@
 ﻿using System;
 
-namespace ChessSharp.Pieces
+namespace ChessSharp.Pieces;
+
+[Flags]
+internal enum PawnMoveType
 {
-    [Flags]
-    internal enum PawnMoveType
-    {
-        Invalid = 0,
-        OneStep = 1,
-        TwoSteps = 2,
-        Capture = 4,
-        Promotion = 8
-    }
+    Invalid = 0,
+    OneStep = 1,
+    TwoSteps = 2,
+    Capture = 4,
+    Promotion = 8
 }
