@@ -1,15 +1,14 @@
-﻿namespace ChessSharp.Pieces
+﻿namespace ChessSharp.Pieces;
+
+/// <summary>Represents the promotion of the pawn.</summary>
+public enum PawnPromotion
 {
-    /// <summary>Represents the promotion of the pawn.</summary>
-    public enum PawnPromotion
-    {
-        /// <summary>Promote the pawn to a knight.</summary>
-        Knight,
-        /// <summary>Promote the pawn to a bishop.</summary>
-        Bishop,
-        /// <summary>Promote the pawn to a rook.</summary>
-        Rook,
-        /// <summary>Promote the pawn to a queen.</summary>
-        Queen,
-    }
+    /// <summary>Promote the pawn to a knight.</summary>
+    Knight,
+    /// <summary>Promote the pawn to a bishop.</summary>
+    Bishop,
+    /// <summary>Promote the pawn to a rook.</summary>
+    Rook,
+    /// <summary>Promote the pawn to a queen.</summary>
+    Queen,
 }
