@@ -1,14 +1,13 @@
-﻿namespace ChessSharp.Pieces
-{
-    /// <summary>Represents a queen <see cref="Piece"/>.</summary>
-    public class Queen : Piece
-    {
-        internal Queen(Player player) : base(player) { }
+﻿namespace ChessSharp.Pieces;
 
-        internal override bool IsValidGameMove(Move move, ChessGame board)
-        {
-            return new Rook(move.Player).IsValidGameMove(move, board) ||
-                   new Bishop(move.Player).IsValidGameMove(move, board);
-        }
+/// <summary>Represents a queen <see cref="Piece"/>.</summary>
+public class Queen : Piece
+{
+    internal Queen(Player player) : base(player) { }
+
+    internal override bool IsValidGameMove(Move move, ChessGame board)
+    {
+        return new Rook(move.Player).IsValidGameMove(move, board) ||
+               new Bishop(move.Player).IsValidGameMove(move, board);
     }
 }
