@@ -1,11 +1,10 @@
-﻿namespace ChessSharp
+﻿namespace ChessSharp;
+
+/// <summary>Represents a player.</summary>
+public enum Player
 {
-    /// <summary>Represents a player.</summary>
-    public enum Player
-    {
-        /// <summary>Represents the white player.</summary>
-        White,
-        /// <summary>Represents the black player.</summary>
-        Black
-    }
+    /// <summary>Represents the white player.</summary>
+    White,
+    /// <summary>Represents the black player.</summary>
+    Black
 }
