@@ -1,24 +1,23 @@
-﻿namespace ChessSharp.SquareData
-{
+﻿namespace ChessSharp.SquareData;
 
-    /// <summary>Represents a square file.</summary>
-    public enum File
-    {
-        /// <summary>Represents the A file.</summary>
-        A,
-        /// <summary>Represents the B file.</summary>
-        B,
-        /// <summary>Represents the C file.</summary>
-        C,
-        /// <summary>Represents the D file.</summary>
-        D,
-        /// <summary>Represents the E file.</summary>
-        E,
-        /// <summary>Represents the F file.</summary>
-        F,
-        /// <summary>Represents the G file.</summary>
-        G,
-        /// <summary>Represents the H file.</summary>
-        H
-    }
+
+/// <summary>Represents a square file.</summary>
+public enum File
+{
+    /// <summary>Represents the A file.</summary>
+    A,
+    /// <summary>Represents the B file.</summary>
+    B,
+    /// <summary>Represents the C file.</summary>
+    C,
+    /// <summary>Represents the D file.</summary>
+    D,
+    /// <summary>Represents the E file.</summary>
+    E,
+    /// <summary>Represents the F file.</summary>
+    F,
+    /// <summary>Represents the G file.</summary>
+    G,
+    /// <summary>Represents the H file.</summary>
+    H
 }
