@@ -1,24 +1,23 @@
-﻿namespace ChessSharp.SquareData
-{
+﻿namespace ChessSharp.SquareData;
 
-    /// <summary>Represents a square rank.</summary>
-    public enum Rank
-    {
-        /// <summary>Represents the first rank.</summary>
-        First,
-        /// <summary>Represents the second rank.</summary>
-        Second,
-        /// <summary>Represents the third rank.</summary>
-        Third,
-        /// <summary>Represents the forth rank.</summary>
-        Forth,
-        /// <summary>Represents the fifth rank.</summary>
-        Fifth,
-        /// <summary>Represents the sixth rank.</summary>
-        Sixth,
-        /// <summary>Represents the seventh rank.</summary>
-        Seventh,
-        /// <summary>Represents the eighth rank.</summary>
-        Eighth
-    }
+
+/// <summary>Represents a square rank.</summary>
+public enum Rank
+{
+    /// <summary>Represents the first rank.</summary>
+    First,
+    /// <summary>Represents the second rank.</summary>
+    Second,
+    /// <summary>Represents the third rank.</summary>
+    Third,
+    /// <summary>Represents the forth rank.</summary>
+    Forth,
+    /// <summary>Represents the fifth rank.</summary>
+    Fifth,
+    /// <summary>Represents the sixth rank.</summary>
+    Sixth,
+    /// <summary>Represents the seventh rank.</summary>
+    Seventh,
+    /// <summary>Represents the eighth rank.</summary>
+    Eighth
 }
