@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace ChessSharp;
 
-namespace ChessSharp
+interface IDeepCloneable<T>
 {
-    interface IDeepCloneable<T>
-    {
-        T DeepClone();
-    }
+    T DeepClone();
 }
