@@ -1,4 +1,1 @@
-﻿global using System.Collections.Immutable;
-global using Microsoft.Extensions.DependencyInjection;
-global using Microsoft.Extensions.Logging;
-global using ApplicationExecutionState = Windows.ApplicationModel.Activation.ApplicationExecutionState;
+﻿global using Microsoft.Extensions.Logging;
